fix exercicie 7

+++ lotofacil/lib.lotofacil/Models/Jogo.cs
@@ -2,6 +2,6 @@ public class Jogo{
     public string Nome {get;set;}
     public string CPF {get;set;}
     public int[] NumerosSelecionados {get;set;}
-    public double Premiacao { get;set; }
+    public decimal Premiacao { get;set; }
 
 }
